Fix "no method error: compile" bug
A compiler engine needs to be instantiated in
ClojurescriptTemplate#prepare, not in #initialize_engine. It seems that
initialize_engine is only called once, but a new instance of the
template class is created for each compile. As such, all compile runs
after the first were failing, because @compiler was never being set.

This could be fixed by making @compiler a class-level var, and using
only one instance of it for all compile passes, but for now I'm just
going to instantiate a new Compiler in #prepare.